requirements.txt
tensorflow~=2.20.0rc0
numpy~=2.3.2
matplotlib~=3.10.5
tensorflow-datasets
Pillow~=11.3.0
keras~=3.11.1
scikit-image~=0.25.2